0.9.0 - Create Catapult Task

+++ include/autonomous.hpp
@@ -1,6 +1,6 @@
 extern void drive(double length);
-extern void drive(double length, int velocity);
+extern void drive(double length, unsigned short int velocity);
 extern void turn(char direction, double length);
-extern void turn(char direction, double length, int velocity);
-extern void AutonIntake(short int msec, bool isReverse);
+extern void turn(char direction, double length, unsigned short int velocity);
+extern void AutonIntake(unsigned short int msec, bool isReverse);
 extern void AutonCatapult();
